lifesync_ai_backend/pyproject.toml
[project]
name = "lifesync-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "pydantic[email]==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "aiofiles==23.2.1",
    "httpx==0.25.2",
    "python-docx==1.1.0",
    "PyPDF2==3.0.1",
    "celery==5.3.4",
    "redis==5.0.1",
    "alembic==1.12.1",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "numpy>=2.3.2",
    "requests>=2.32.4",
    "node>=1.2.2",
    "npm>=0.1.1",
    "bcrypt>=4.3.0",
]
